requirements.txt
# LangChainSkills 核心依赖
langchain>=1.0.0
langchain-deepseek>=0.1.0
langgraph>=0.1.0
openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# FastAPI 服务依赖
fastapi>=0.100.0
uvicorn>=0.23.0
aiofiles>=23.0.0
prometheus-client>=0.16.0

# Skill 依赖（数据分析）
numpy>=1.24.0
matplotlib>=3.7.0

# Skill 依赖（PDF 处理）
pdfplumber>=0.7.0
pandas>=2.0.0
